+++ theme/components/Bio.vue
@@ -63,7 +63,7 @@
       :subtitle-text="presenterData.title"
     >
       <ui5-avatar slot="avatar" :initials="presenterData.initials" size="M" shape="Circle">
-        <img v-if="presenterData.photo" :src="presenterData.photo" :alt="presenterData.name" />
+        <img v-if="presenterData.photo" :src="assetUrl(presenterData.photo)" :alt="presenterData.name" />
       </ui5-avatar>
     </ui5-card-header>
     <!-- eslint-enable vue/no-deprecated-slot-attribute -->